rename eas-core to ddd-core than re-organized it
