+++ ddd-core/pom.xml
@@ -10,7 +10,7 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>eas-core</artifactId>
+    <artifactId>ddd-core</artifactId>
 
     <dependencies>
         <dependency>
