+++ ddd-core/src/main/java/xyz/zhangyi/ddd/core/domain/Identity.java
@@ -1,4 +1,4 @@
-package xyz.zhangyi.ddd.eas.core.domain;
+package xyz.zhangyi.ddd.core.domain;
 
 public interface Identity<T> {
     T value();
