+++ ddd-core/src/main/java/xyz/zhangyi/ddd/core/exception/ApplicationInfrastructureException.java
@@ -1,6 +1,4 @@
-package xyz.zhangyi.ddd.eas.core.exception;
-
-import xyz.zhangyi.ddd.eas.core.exception.ApplicationException;
+package xyz.zhangyi.ddd.core.exception;
 
 public class ApplicationInfrastructureException extends ApplicationException {
     public ApplicationInfrastructureException(String message, Exception ex) {
