+++ ddd-core/src/main/java/xyz/zhangyi/ddd/core/exception/ApplicationValidationException.java
@@ -1,6 +1,4 @@
-package xyz.zhangyi.ddd.eas.core.exception;
-
-import xyz.zhangyi.ddd.eas.core.exception.ApplicationException;
+package xyz.zhangyi.ddd.core.exception;
 
 public class ApplicationValidationException extends ApplicationException {
     public ApplicationValidationException(String message) {
