+++ ddd-core/src/main/java/xyz/zhangyi/ddd/core/gateway/acl/port/EventPublisher.java
@@ -1,6 +1,6 @@
-package xyz.zhangyi.ddd.eas.core.gateway.acl.port;
+package xyz.zhangyi.ddd.core.gateway.acl.port;
 
-import xyz.zhangyi.ddd.eas.core.event.Event;
+import xyz.zhangyi.ddd.core.event.Event;
 
 public interface EventPublisher<T extends Event> {
     void publish(T event);
