+++ eas-employee/pom.xml
@@ -23,7 +23,7 @@
         </dependency>
         <dependency>
             <groupId>xyz.zhangyi.ddd</groupId>
-            <artifactId>eas-core</artifactId>
+            <artifactId>ddd-core</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
     </dependencies>
