+++ eas-employee/src/main/java/xyz/zhangyi/ddd/eas/valueadded/employeecontext/domain/exception/InvalidAttendanceException.java
@@ -1,6 +1,6 @@
 package xyz.zhangyi.ddd.eas.valueadded.employeecontext.domain.exception;
 
-import xyz.zhangyi.ddd.eas.core.domain.exception.DomainException;
+import xyz.zhangyi.ddd.core.exception.DomainException;
 
 public class InvalidAttendanceException extends DomainException {
     public InvalidAttendanceException(String message) {
