+++ eas-project/src/main/java/xyz/zhangyi/ddd/eas/projectcontext/domain/exception/AssignmentIssueException.java
@@ -1,6 +1,6 @@
 package xyz.zhangyi.ddd.eas.projectcontext.domain.exception;
 
-import xyz.zhangyi.ddd.eas.core.domain.exception.DomainException;
+import xyz.zhangyi.ddd.core.exception.DomainException;
 
 public class AssignmentIssueException extends DomainException {
     public AssignmentIssueException(String message) {
