+++ eas-project/src/main/java/xyz/zhangyi/ddd/eas/projectcontext/domain/exception/IssueException.java
@@ -1,6 +1,6 @@
 package xyz.zhangyi.ddd.eas.projectcontext.domain.exception;
 
-import xyz.zhangyi.ddd.eas.core.domain.exception.DomainException;
+import xyz.zhangyi.ddd.core.exception.DomainException;
 
 public class IssueException extends DomainException {
     public IssueException(String message) {
