+++ eas-training/pom.xml
@@ -20,7 +20,7 @@
         </dependency>
         <dependency>
             <groupId>xyz.zhangyi.ddd</groupId>
-            <artifactId>eas-core</artifactId>
+            <artifactId>ddd-core</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
