+++ eas-training/src/main/java/xyz/zhangyi/ddd/eas/valueadded/trainingcontext/domain/exception/NominationException.java
@@ -1,6 +1,6 @@
 package xyz.zhangyi.ddd.eas.valueadded.trainingcontext.domain.exception;
 
-import xyz.zhangyi.ddd.eas.core.domain.exception.DomainException;
+import xyz.zhangyi.ddd.core.exception.DomainException;
 
 public class NominationException extends DomainException {
     public NominationException(String message) {
