+++ eas-training/src/main/java/xyz/zhangyi/ddd/eas/valueadded/trainingcontext/domain/notification/MailTemplateException.java
@@ -1,6 +1,6 @@
 package xyz.zhangyi.ddd.eas.valueadded.trainingcontext.domain.notification;
 
-import xyz.zhangyi.ddd.eas.core.domain.exception.DomainException;
+import xyz.zhangyi.ddd.core.exception.DomainException;
 
 public class MailTemplateException extends DomainException {
     public MailTemplateException(String message) {
