+++ eas-training/src/main/java/xyz/zhangyi/ddd/eas/valueadded/trainingcontext/domain/validdate/ValidDateException.java
@@ -1,6 +1,6 @@
 package xyz.zhangyi.ddd.eas.valueadded.trainingcontext.domain.validdate;
 
-import xyz.zhangyi.ddd.eas.core.domain.exception.DomainException;
+import xyz.zhangyi.ddd.core.exception.DomainException;
 
 public class ValidDateException extends DomainException {
     public ValidDateException(String message) {
