+++ eas-training/src/main/java/xyz/zhangyi/ddd/eas/valueadded/trainingcontext/ohs/remote/resource/TrainingResource.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import xyz.zhangyi.ddd.eas.core.gateway.ohs.Resources;
+import xyz.zhangyi.ddd.core.gateway.ohs.Resources;
 import xyz.zhangyi.ddd.eas.valueadded.trainingcontext.ohs.local.appservice.TrainingAppService;
 import xyz.zhangyi.ddd.eas.valueadded.trainingcontext.ohs.local.pl.TrainingResponse;
